+++ packages/sp-proxy/src/frameworks-drivers/main/protocols/BooleanStringType.ts
@@ -1,0 +1,7 @@
+export type BooleanStringType =
+  | 'TRUE'
+  | 'true'
+  | 'True'
+  | 'FALSE'
+  | 'false'
+  | 'False'