+++ packages/sp-proxy/src/frameworks-drivers/main/protocols/IStringParser.ts
@@ -1,0 +1,6 @@
+import { BooleanStringType } from './BooleanStringType'
+
+export interface IStringParser {
+  stringToInt: (str: string) => number
+  stringToBool: (booleanString: BooleanStringType) => boolean
+}