+++ packages/sp-proxy/src/frameworks-drivers/main/server.ts
@@ -1,16 +1,27 @@
+import cfg from '@sp-proxy/frameworks-drivers/main/config/env'
 import routes from '@sp-proxy/frameworks-drivers/main/routes'
 import express from 'express'
 import { readFileSync } from 'fs'
+import http from 'http'
 import https from 'https'
-import cfg from '@sp-proxy/frameworks-drivers/main/config/env'
+import { WinstonLogger } from './logger/WinstonLogger'
 
 const app = express()
-
+const logger = WinstonLogger.getInstance()
 app.use(routes)
 app.use(express.urlencoded({ extended: false }))
-const cert = readFileSync(cfg.tlsCertPath).toString()
-const key = readFileSync(cfg.tlsKeyPath).toString()
-const server = https.createServer({ key: key, cert: cert }, app)
-server.listen(cfg.port, () => console.log('server started'))
+if (!cfg.useTls) {
+  const server = http.createServer(app)
+  server.listen(cfg.port, () =>
+    logger.info('Inbound Saml SP Proxy Started in non TLS mode')
+  )
+} else {
+  const cert = readFileSync(cfg.tlsCertPath).toString()
+  const key = readFileSync(cfg.tlsKeyPath).toString()
+  const server = https.createServer({ key: key, cert: cert }, app)
+  server.listen(cfg.port, () =>
+    logger.info('Inbound Saml SP Proxy Started in TLS mode')
+  )
+}
 
 export default app